+++ project/.gitignore
@@ -1,4 +1,4 @@
-env/
+../env/
 venv/
 log.log
 .idea/
