+++ project/requirements.txt
@@ -1,0 +1,21 @@
+attrs==21.4.0
+certifi==2022.5.18.1
+charset-normalizer==2.0.12
+Faker==13.12.1
+idna==3.3
+iniconfig==1.1.1
+jsonschema==4.6.1
+packaging==21.3
+pip==22.1.2
+pluggy==1.0.0
+py==1.11.0
+pyparsing==3.0.9
+pyrsistent==0.18.1
+pytest==7.1.2
+python-dateutil==2.8.2
+requests==2.27.1
+setuptools==60.2.0
+six==1.16.0
+tomli==2.0.1
+urllib3==1.26.9
+wheel==0.37.1
