+++ test_module/test_path.py
@@ -1,6 +1,6 @@
 import json
 import pytest
-import main
+from project.utils import create_fake_dict, create_output_file, existing_dir, clear_files_in_path, construct_files
 import os
 
 @staticmethod
@@ -17,7 +17,7 @@ def test_dict_len_matches_key_number():
     data_schema = "{\"date\": \"timestamp:\",\"name\": \"str:rand\",\"type\": \"['client', 'partner', 'government']\"," \
                   "\"animal_type\": \"['cat', 'dog', 'monkey','tiger']\",\"age\": \"int:rand(1, 90)\",\"kids_number\": " \
                   "\"int:rand(1, 6)\"} "
-    dict_given = main.create_fake_dict(data_schema)
+    dict_given = create_fake_dict(data_schema)
     dict_values = dict_given.values()
     dict_keys = dict_given.keys()
     assert len(dict_values) == len(dict_keys)
@@ -27,7 +27,7 @@ def test_add_lines_to_file():
     d_schema = "{\"date\": \"timestamp:\"} "
     path = "/Users/kpodlaska/Desktop"
     f_line_given = 10
-    file_given = main.create_output_file(f_line_given, d_schema, path, file_name)
+    file_given = create_output_file(f_line_given, d_schema, path, file_name)
     print(file_given)
     f = open(file_given)
     data = json.load(f)
@@ -36,7 +36,7 @@ def test_add_lines_to_file():
 
 @staticmethod
 def count_files_in_path(path, file_name):
-    main.existing_dir(path)
+    existing_dir(path)
     "check if path exists"
     files = os.listdir(path)
     result = []
@@ -53,8 +53,8 @@ def test_creating_number_of_files():
     d_schema = "{\"date\": \"timestamp:\"} "
     f_name = "test_file_for_testing_path.json"
     path= "/Users/kpodlaska/Desktop"
-    cleaning_before_testing = main.clear_files_in_path(path, f_name)
-    main.create_output_file(20, d_schema, path, f_name)
+    cleaning_before_testing = clear_files_in_path(path, f_name)
+    create_output_file(20, d_schema, path, f_name)
     assert count_files_in_path(path, f_name) == expected_value
 
 
@@ -66,11 +66,11 @@ def test_clear_path():
         f_name = "test_file_for_testing_path.json"
         prefix="count"
         path = "/Users/kpodlaska/Desktop"
-        cleaning_before_testing = main.clear_files_in_path(path,f_name)
+        cleaning_before_testing = clear_files_in_path(path, f_name)
 
-        files = main.construct_files(f_name,prefix, how_many_files)
+        files = construct_files(f_name, prefix, how_many_files)
         for file in files:
-            created_file = main.create_output_file(f_line,d_schema,path,file)
+            created_file = create_output_file(f_line, d_schema, path, file)
         assert count_files_in_path(path, f_name) == how_many_files
 
 
@@ -90,7 +90,7 @@ testdata = [("{\"age\": \"int:rand(1:90)\"} ", True),("{\"kids_number\": \"int:r
 
 @pytest.mark.parametrize("a, expected", testdata)
 def test_is_digit(a, expected):
-    values = list(main.create_fake_dict(a).values())
+    values = list(create_fake_dict(a).values())
     given = values[0]
     assert (int(given)/1).is_integer() == expected
 
@@ -101,6 +101,5 @@ test_schema = [("{\"date\": \"timestamp:\",\"name\": \"str:rand\",\"type\": \"['
 def test_data_schema_is_json_format(given, expected):
     assert validateJsonText(given) == expected
 
-if __name__ == '__main__':
-    main()
+
